requirements.txt
aiohttp_jinja2
cchardet
pyyaml
pytest-aiohttp
pep8
pep257
